+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - audioRead function to convert more audio file types to Wave object
 - hardRain package implementation of rainfall detection
+- parseFilename n ow accepts POSIX timestamp as a format
 
 # sonicscrewdriver 0.0.3